==> 03) Data Prep.ipynb <==
import gensim
import pandas as pd 
import numpy as np 
import time
cd = 'C:\\Users\\Giada\\Documents\\GitHub\\Tweet-Sentiment-Predictor\\'

### Creating a list of categories for each word in the vocab to be categorised into depending on its similarity to other words


#If two words have a similarity of more than the limit then this function removes the less frequent word from the list
def similar_word_remover(model,word_list,similarity_limit):
    new_list = []
    for i in range(len(word_list)):
        for j in range(len(word_list)):
            if model.wv.similarity(word_list[i],word_list[j]) > similarity_limit and model.wv.similarity(word_list[i],word_list[j]) != 1:
                if i>j:
                    new_list.append(word_list[i])
                else:
                    new_list.append(word_list[j])
    new_list = list(dict.fromkeys(new_list))

    final_list = word_list
    for word in new_list:
        final_list.remove(word)
    return final_list

#Function to create a similarity matrix in a dataframe and output it to a csv file
def similarity_matrix(model, word_list):
    df = pd.DataFrame(columns = word_list, index=word_list)
    for i in range(len(word_list)):
        for j in range(len(word_list)):
            df.iloc[i,j] = w2vmodel.wv.similarity(word_list[i],word_list[j])
    df.to_csv('similarity_matrix_data.csv',index=True)


#How many of the most frequent words do you want the categories to be chosen from?
number_of_catagories = 250
#Upper limit for word similarity
similarity_limit = 0.75

w2vmodel = gensim.models.KeyedVectors.load('w2vmodel')

#Loading all the values form the word2vec into a dictionary and sorting them based on their frequency
w2c = dict()
for item in w2vmodel.wv.vocab:
    w2c[item]=w2vmodel.wv.vocab[item].count
w2cSorted=dict(sorted(w2c.items(), key = lambda x : x[1],reverse=True))

#Creating a list of the most frequent words
dict_words = w2cSorted.keys()
ordered_words = list(dict_words)
most_frequent_words = ordered_words[0:number_of_catagories]

new_word_list = similar_word_remover(model=w2vmodel, word_list=most_frequent_words,similarity_limit=similarity_limit)
similarity_matrix(model=w2vmodel, word_list=new_word_list)

#Writing the final dataset categories to a txt file
with open('Dataset_Categories.txt','w') as file:
    for word in new_word_list:
        file.write(word + "\n")
    file.write("OTHER" + "\n")
file.close()

### Running through the tweets and creating a dataset of frequency for each 

#Loops each word through the list of categories and assigns the word the most similar category
def category_chooser(word,categories,w2vmodel):
    most_similar_category = len(categories)-1
    #best similarity chosen as a minimum cosine similarity needed for another category to be chosen
    best_similarity = 0.6
    for i in range(len(categories)-1):
        temp_similarity = w2vmodel.wv.similarity(word,categories[i])
        if temp_similarity > best_similarity: 
            best_similarity = temp_similarity
            most_similar_category = i
    return most_similar_category

#THIS CODE RUNS THROUGH EACH WORD OF EACH TWEET AND ASSIGNS IT TO A CATEGORY
#THE CODE TAKES A LONG TIME TO RUN: ~45mins

start_time = time.time()
#Reading a long list of categories from the dataset catagory file
with open('Dataset_Categories.txt','r') as file:
    categories = file.read().splitlines()
file.close()

df = pd.read_csv('tweets_Combined.csv', low_memory=False)
SCORE = df.pop("SCORE")
dataset_final = pd.DataFrame(columns=categories,index=range(len(df)))
w2vmodel = gensim.models.KeyedVectors.load('w2vmodel')

#Loop to go through each of the tweets
for tweet_number in range(round(len(df)/1000)): 

    #Initial setting of all categories to 0 
    dataset_final.iloc[tweet_number,:] = 0

    #Loop through all of the words in the tweet
    for word_number in range(50):
        word = df.iloc[tweet_number,word_number]
        #If statement to check the words are in the model vocab
        if word is not np.nan and word in w2vmodel.wv.vocab.keys():
            category_number = category_chooser(word=word,categories=categories,w2vmodel=w2vmodel)
            dataset_final.iloc[tweet_number,category_number] = dataset_final.iloc[tweet_number,category_number] + 1

print("Time taken: %.2f" % (time.time() - start_time))
dataset_final["SCORE"] = SCORE

dataset_final.to_csv('modelling_dset.csv',index=True)
